Remove 'arguments' from commands.py
